In vldcp(4), initialize lc_rx_state to a magic value which is not used by the hypervisor interface. This allows the driver to sync up with the firmware's Rx channel state (DOWN/UP/RESET) upon the first interrupt, and repair inconsistent Rx channel head/tail queue state if necessary.
Unwedges another interrupt storm observed when starting ldomd on my
T5220 after a reboot. In case the firmware's Rx queue head and tail
are initially not equal, the normal interrupt handling path will not
operate as expected due to the interrupt storm, and we never told the
firmware to shut up.

Other cbus(4) drivers might benefit from a similar fix, but we're
too close to release for experiments, so I'm just fixing vldcp(4).

ok kettenis@ deraadt@