Insert definitions of environ and __progname in MD_CRT0_START again.
That part is not needed with a PIE toolchain, but until gcc 4 on m88k
receives enough fixes to produce as good code as gcc 3, the bits in
r1.5 are still required.

This fixes miniperl does not work correctly while building perl on m88k.

Diff is suggested by Miod Vallat, tested by him and me, ok guenther@